chore: move updater files to subdirectory
